Further consolidation of namespaces
